move configItem & textinputconfig to new file
