+++ app/config/signupForm.ts
@@ -3,15 +3,7 @@ import {
   validateMinLength,
   validateName,
 } from '../utils/formValidation';
-import type { FormValues, ValidationFunction } from '../types/signupForm';
-
-interface TextInputConfig {
-  field: keyof FormValues;
-  label: string;
-  placeholder: string;
-  validate?: ValidationFunction;
-  required?: boolean;
-}
+import type { FormValues, TextInputConfig, ValidationFunction } from '../types/signupForm';
 
 /**
  * Creates a text input configuration object for form fields.
